+++ src/main/kotlin/days/Day4.kt
@@ -3,7 +3,7 @@ package days
 import Solution
 
 class Day4 : Solution {
-	override val day: Int = 4
+	override val day: UByte = 4u
 
 	override fun solve1(): String {
 		var sum: Int = 0
